+++ rest-api-spec/src/main/resources/rest-api-spec/test/cluster.stats/10_basic.yml
@@ -83,3 +83,86 @@
       cluster.stats: {}
 
   - is_true: nodes.packaging_types
+---
+"get cluster stats without runtime fields":
+  - skip:
+      version: " - 7.99.99"
+      reason:  "cluster stats includes runtime fields from 8.0 on"
+  - do:
+      indices.create:
+        index: sensor
+
+  - do: {cluster.stats: {}}
+  - length: { indices.mappings.field_types: 0 }
+  - length: { indices.mappings.runtime_field_types: 0 }
+
+---
+"Usage stats with script-less runtime fields":
+  - skip:
+      version: " - 7.99.99"
+      reason:  "cluster stats includes runtime fields from 8.0 on"
+  - do:
+      indices.create:
+        index: sensor
+        body:
+          mappings:
+            runtime:
+              message_from_source:
+                type: keyword
+              bad_map:
+                type: double    # shadows the bad_map field in properties
+              message.text:
+                type: keyword   # shadows the message.text subfield in properties
+            properties:
+              message:
+                type: keyword
+                fields:
+                  text:
+                    type: text
+              bad_map:
+                type: long
+
+  - do: {cluster.stats: {}}
+  - length: { indices.mappings.field_types: 3 }
+
+  - match: { indices.mappings.field_types.0.name: keyword }
+  - match: { indices.mappings.field_types.0.count: 1 }
+  - match: { indices.mappings.field_types.0.index_count: 1 }
+  - match: { indices.mappings.field_types.1.name: long }
+  - match: { indices.mappings.field_types.1.count: 1 }
+  - match: { indices.mappings.field_types.1.index_count: 1 }
+  - match: { indices.mappings.field_types.2.name: text }
+  - match: { indices.mappings.field_types.2.count: 1 }
+  - match: { indices.mappings.field_types.2.index_count: 1 }
+
+
+  - length: { indices.mappings.runtime_field_types: 2 }
+
+  - match: { indices.mappings.runtime_field_types.0.name: double }
+  - match: { indices.mappings.runtime_field_types.0.count: 1 }
+  - match: { indices.mappings.runtime_field_types.0.index_count: 1 }
+  - match: { indices.mappings.runtime_field_types.0.scriptless_count: 1 }
+  - match: { indices.mappings.runtime_field_types.0.shadowed_count: 1 }
+  - match: { indices.mappings.runtime_field_types.0.source_max: 0 }
+  - match: { indices.mappings.runtime_field_types.0.source_total: 0 }
+  - match: { indices.mappings.runtime_field_types.0.lines_max:  0 }
+  - match: { indices.mappings.runtime_field_types.0.lines_total: 0 }
+  - match: { indices.mappings.runtime_field_types.0.chars_max:  0 }
+  - match: { indices.mappings.runtime_field_types.0.chars_total: 0 }
+  - match: { indices.mappings.runtime_field_types.0.doc_max: 0 }
+  - match: { indices.mappings.runtime_field_types.0.doc_total: 0 }
+
+  - match: { indices.mappings.runtime_field_types.1.name: keyword }
+  - match: { indices.mappings.runtime_field_types.1.count: 2 }
+  - match: { indices.mappings.runtime_field_types.1.index_count: 1 }
+  - match: { indices.mappings.runtime_field_types.1.scriptless_count: 2 }
+  - match: { indices.mappings.runtime_field_types.1.shadowed_count: 1 }
+  - match: { indices.mappings.runtime_field_types.1.source_max: 0 }
+  - match: { indices.mappings.runtime_field_types.1.source_total: 0 }
+  - match: { indices.mappings.runtime_field_types.1.lines_max:  0 }
+  - match: { indices.mappings.runtime_field_types.1.lines_total: 0 }
+  - match: { indices.mappings.runtime_field_types.1.chars_max:  0 }
+  - match: { indices.mappings.runtime_field_types.1.chars_total: 0 }
+  - match: { indices.mappings.runtime_field_types.1.doc_max: 0 }
+  - match: { indices.mappings.runtime_field_types.1.doc_total: 0 }
+
