+++ x-pack/plugin/core/src/main/java/org/elasticsearch/xpack/core/XPackField.java
@@ -68,8 +68,6 @@ public final class XPackField {
     public static final String DATA_TIERS = "data_tiers";
     /** Name constant for the aggregate_metric plugin. */
     public static final String AGGREGATE_METRIC = "aggregate_metric";
-    /** Name constant for the runtime fields plugin. */
-    public static final String RUNTIME_FIELDS = "runtime_fields";
     /** Name constant for the operator privileges feature. */
     public static final String OPERATOR_PRIVILEGES = "operator_privileges";
