--- x-pack/plugin/runtime-fields/build.gradle
@@ -1,21 +1,0 @@
-apply plugin: 'elasticsearch.esplugin'
-
-esplugin {
-  name 'x-pack-runtime-fields'
-  description 'A module which adds support for runtime fields'
-  classname 'org.elasticsearch.xpack.runtimefields.RuntimeFields'
-  extendedPlugins = ['x-pack-core']
-}
-archivesBaseName = 'x-pack-runtime-fields'
-
-dependencies {
-  compileOnly project(":server")
-  compileOnly project(path: xpackModule('core'))
-}
-
-tasks.named("dependencyLicenses").configure {
-  ignoreSha 'x-pack-core'
-}
-
-//this plugin is here only for the painless extension, there are no unit tests
-tasks.named("testingConventions").configure { enabled = false }
