+++ x-pack/plugin/security/qa/operator-privileges-tests/src/javaRestTest/java/org/elasticsearch/xpack/security/operator/Constants.java
@@ -254,7 +254,6 @@ public class Constants {
         "cluster:monitor/xpack/info/ml",
         "cluster:monitor/xpack/info/monitoring",
         "cluster:monitor/xpack/info/rollup",
-        "cluster:monitor/xpack/info/runtime_fields",
         "cluster:monitor/xpack/info/searchable_snapshots",
         "cluster:monitor/xpack/info/security",
         "cluster:monitor/xpack/info/slm",
@@ -307,7 +306,6 @@ public class Constants {
         "cluster:monitor/xpack/usage/ml",
         "cluster:monitor/xpack/usage/monitoring",
         "cluster:monitor/xpack/usage/rollup",
-        "cluster:monitor/xpack/usage/runtime_fields",
         "cluster:monitor/xpack/usage/searchable_snapshots",
         "cluster:monitor/xpack/usage/security",
         "cluster:monitor/xpack/usage/slm",
